Fix concurrent access
The underlying Map is not thread safe and is already protected by a
`synchronized` in `getConfiguration(String)` so a similar protection is
needed when reading the whole map content in `copyCachedState`.
To do so a synchronized `copyCachedConfigurations` has been added.
In its call, the calling instance can access its own configurations
safely because it is under construction.